requirements.txt
django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2-binary>=2.9
gunicorn==23.0.0
django-cors-headers==4.6.0
python-dotenv==1.0.1
requests>=2.31.0
drf-spectacular==0.28.0
